Backend/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
python-multipart
itsdangerous
